move log_stick_expectation to cython
