+++ bnp/__init__.py
@@ -1,7 +1,1 @@
-from .online_hdp import HierarchicalDirichletProcess
-
 __version__ = '0.0.1'
-
-__all__ = [
-    'HierarchicalDirichletProcess'
-]
