+++ bnp/utils/expectation.py
@@ -31,36 +31,6 @@ def log_dirichlet_expectation(alpha):
         return (psi(alpha) - psi(np.sum(alpha, 1))[:, np.newaxis])
 
 
-def log_stick_expectation(sticks):
-    """Expectation of log stick-breaking process
-
-    Parameters
-    ----------
-    sticks : array, [2 ,k]
-        Each column is a pair of parameter of Beta distribution a_{k}, b_{k}
-
-    Returns
-    -------
-    Elogsticks: array, [k+1,]
-        this is E[log(sticks)]
-
-    """
-
-    # psi(a_{k} + b_{k}) for k = {1,2,..., K-1}
-    stick_sum = psi(np.sum(sticks, 0))
-    # E[log(V_{k})] = psi(a_{k}) - psi(a_{k} + b_{k}) for k = {1,2,...,K-1}
-    expectation_log_v = psi(sticks[0]) - stick_sum
-    # E[log(1 - V_{k})] = psi(b_{k}) - psi(a_{k} + b_{k}) for k = {1,2,...,K-1}
-    expacetaion_log_1_minus_v = psi(sticks[1]) - stick_sum
-
-    size = sticks.shape[1] + 1
-    Elogsticks = np.zeros(size)
-    # E[log(sigma_{k}(V))] = E[log(V_{k})] + sum_{1 to k-1}(E[log(1 - V_{l})])
-    Elogsticks[0: (size - 1)] = expectation_log_v
-    Elogsticks[1:] = Elogsticks[1:] + np.cumsum(expacetaion_log_1_minus_v)
-    return Elogsticks
-
-
 def stick_expectation(sticks):
     """Expectation of stick-breaking process
