+++ requirements.txt
@@ -1,3 +1,5 @@
 numpy>=1.11.2
 scipy>=0.18.1
 scikit-learn>=0.18.1
+Cython>=0.20.2
+
